% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNCode1.R
\docType{data}
\name{Packages}
\alias{Packages}
\title{First 6 Line  Code Testing}
\format{An object of class \code{character} of length 8.}
\usage{
Packages
}
\value{
data
}
\description{
First 6 Line  Code Testing
}
\keyword{datasets}
